# DMNLiteralExpressionEvaluator to internalize FEEL