main port of https://github.com/kiegroup/drools/pull/4048

like:
 - DMNInvocationEvaluator
 - DMNDTExpressionEvaluator

BEFORE
Benchmark
(numberOfDecisionsWithContext)  Mode    Cnt  Score   Error  Units
DMNEvaluateTriangularNumHardBenchmark.evaluateContext
20    ss  50000  0.103 ± 0.001  ms/op
DMNEvaluateTriangularNumHardBenchmark.evaluateContext
50    ss  50000  0.245 ± 0.001  ms/op
AFTER
Benchmark
(numberOfDecisionsWithContext)  Mode    Cnt  Score   Error  Units
DMNEvaluateTriangularNumHardBenchmark.evaluateContext
20    ss  50000  0.038 ± 0.001  ms/op
DMNEvaluateTriangularNumHardBenchmark.evaluateContext
50    ss  50000  0.141 ± 0.001  ms/op

BEFORE
Benchmark
(numberOfDecisions)  Mode    Cnt  Score   Error  Units
DMNEvaluateFewLiteralDecisionBenchmark.evaluateDecision
20    ss  10000  0.123 ± 0.005  ms/op
DMNEvaluateFewLiteralDecisionBenchmark.evaluateDecision
50    ss  10000  0.297 ± 0.006  ms/op
DMNEvaluateFewLiteralDecisionBenchmark.evaluateDecision
100    ss  10000  0.712 ± 0.006  ms/op
AFTER
Benchmark
(numberOfDecisions)  Mode    Cnt  Score   Error  Units
DMNEvaluateFewLiteralDecisionBenchmark.evaluateDecision
20    ss  10000  0.061 ± 0.001  ms/op
DMNEvaluateFewLiteralDecisionBenchmark.evaluateDecision
50    ss  10000  0.150 ± 0.004  ms/op
DMNEvaluateFewLiteralDecisionBenchmark.evaluateDecision
100    ss  10000  0.503 ± 0.005  ms/op

**Thank you for submitting this pull request**

**NOTE!:** Double check the target branch for this PR.
The default is `main` so it will target Drools 8 / Kogito / Optaplanner 8.
If this PR is not strictly related to drools and kogito project in `drools.git`, it should probably target `7.x`as a branch

**Ports** If a forward-port or a backport is needed, paste the forward port PR here

[link](https://www.example.com)

**JIRA**: _(please edit the JIRA link if it exists)_

[link](https://www.example.com)

**referenced Pull Requests**: _(please edit the URLs of referenced pullrequests if they exist)_

* paste the link(s) from GitHub here
* link 2
* link 3 etc.

<details>
<summary>
How to replicate CI configuration locally?
</summary>

Build Chain tool does "simple" maven build(s), the builds are just Maven commands, but because the repositories relates and depends on each other and any change in API or class method could affect several of those repositories there is a need to use [build-chain tool](https://github.com/kiegroup/github-action-build-chain) to handle cross repository builds and be sure that we always use latest version of the code for each repository.
 
[build-chain tool](https://github.com/kiegroup/github-action-build-chain) is a build tool which can be used on command line locally or in Github Actions workflow(s), in case you need to change multiple repositories and send multiple dependent pull requests related with a change you can easily reproduce the same build by executing it on Github hosted environment or locally in your development environment. See [local execution](https://github.com/kiegroup/github-action-build-chain#local-execution) details to get more information about it.
</details>

<details>
<summary>
How to retest this PR or trigger a specific build:
</summary>

* <b>a pull request</b> please add comment: <b>Jenkins retest this</b>

* <b>a full downstream build</b> please add comment: <b>Jenkins run fdb</b>

* <b>a compile downstream build</b> please  add comment: <b>Jenkins run cdb</b>

* <b>a full production downstream build</b> please add comment: <b>Jenkins execute product fdb</b>

* <b>an upstream build</b> please add comment: <b>Jenkins run upstream</b>
</details>